fix(build,time,trace): include equivalence.c in libasx, widen timer generation, and cap trace event count
Makefile: Add equivalence.c to RUNTIME_SRC so it links into libasx.a,
removing the special-case compile rule for test_codec_equivalence.

timer_wheel.h: Widen timer_handle generation from uint16_t to uint32_t
to avoid rapid wraparound under high timer churn.

trace.c: Cap asx_trace_event_count() return value at ASX_TRACE_CAPACITY
to prevent callers from iterating past readable ring buffer entries when
total emitted events exceed capacity.

+++ include/asx/time/timer_wheel.h
@@ -36,7 +36,7 @@ extern "C" {
 
 typedef struct {
     uint32_t slot;        /* arena slot index */
-    uint16_t generation;  /* generation at registration time */
+    uint32_t generation;  /* generation at registration time */
 } asx_timer_handle;
 
 /* -------------------------------------------------------------------
